General fixes to get things 'v1-ready' (#1)
