+++ README.md
@@ -3,5 +3,7 @@
 This starter is intended to offer a richer starting point as a starter for those looking to create a personal dev or portfolio site.
 
 ## Technologies
-1. Gatsby
-2. TailwindCSS
+1. TypeScript
+2. Gatsby
+3. TailwindCSS
+4. Redux
