+++ studio/schemas/documents/category.ts
